--- app/api/latest/const.ts
@@ -1,1 +1,0 @@
-export const numLatest = 25;
