+++ app/main/play/page.tsx
@@ -1,6 +1,5 @@
 import { ChartBrief } from "@/chartFormat/chart";
 import PlayTab from "./clientPage";
-import { getBrief } from "@/api/brief/brief";
 import { originalCId, sampleCId } from "../const";
 
 export const dynamic = "force-static";
