+++ chartFormat/apiConfig.ts
@@ -1,0 +1,1 @@
+export const rateLimitMin = 10;
