+++ chartFormat/command.ts
