+++ chartFormat/difficulty.ts
