+++ chartFormat/legacy/chart2.ts
