+++ chartFormat/legacy/chart3.ts
