+++ chartFormat/legacy/chart6.ts
