+++ route/api/latest.ts
@@ -1,0 +1,34 @@
+import { Hono } from "hono";
+import { MongoClient } from "mongodb";
+import "dotenv/config";
+import { Bindings } from "../env";
+
+export const numLatest = 25;
+
+const latestApp = new Hono<{ Bindings: Bindings }>({ strict: false }).get(
+  "/",
+  async (c) => {
+    const client = new MongoClient(c.env.MONGODB_URI);
+    try {
+      await client.connect();
+      const db = client.db("nikochan");
+      return c.json(
+        await db
+          .collection("chart")
+          .find({ published: true })
+          .sort({ updatedAt: -1 })
+          .limit(numLatest)
+          .project({ _id: 0, cid: 1 })
+          .toArray(),
+        200,
+        {
+          "cache-control": "max-age=600",
+        }
+      );
+    } finally {
+      await client.close();
+    }
+  }
+);
+
+export default latestApp;
