+++ route/app.ts
@@ -1,6 +1,10 @@
 import { Hono } from "hono";
 import apiApp from "./api/app";
+import { Bindings } from "./env";
 
-const app = new Hono().route("/api", apiApp);
+const app = new Hono<{ Bindings: Bindings }>().route(
+  "/api",
+  apiApp
+);
 
 export default app;
