+++ route/env.ts
@@ -1,0 +1,3 @@
+export interface Bindings {
+  MONGODB_URI: string;
+}
