+++ route/local/index.ts
@@ -1,5 +1,6 @@
 import { serve } from "@hono/node-server";
 import app from "../app";
+import "dotenv/config";
 
 const port = 8787;
 console.log(`Server is running on http://localhost:${port}`);
