+++ tsconfig.json
@@ -23,9 +23,7 @@
       }
     ],
     "paths": {
-      "@/*": [
-        "./app/*"
-      ]
+      "@/*": ["./app/*", "./*"],
     },
     "target": "ES2017"
   },
